requirements.txt
opencv-python
numpy
pyzbar
Pillow
watchdog
ImageHash